Moved SCC computation into own class - (Task #58)
---
Task #58: Numerical stability issue in SteadyStateAvailability iteration
