+++ de.dlr.sc.virsat.fdir.core.test/src/de/dlr/sc/virsat/fdir/core/test/AllTests.java
@@ -17,6 +17,7 @@ import de.dlr.sc.virsat.fdir.core.markov.MarkovAutomatonBuilderTest;
 import de.dlr.sc.virsat.fdir.core.markov.MarkovAutomatonTest;
 import de.dlr.sc.virsat.fdir.core.markov.MarkovTransitionTest;
 import de.dlr.sc.virsat.fdir.core.markov.StronglyConnectedComponentTest;
+import de.dlr.sc.virsat.fdir.core.markov.algorithm.StronglyConnectedComponentFinderTest;
 import de.dlr.sc.virsat.fdir.core.markov.modelchecker.MarkovModelCheckerTest;
 import de.dlr.sc.virsat.fdir.core.markov.modelchecker.ModelCheckingResultTest;
 import de.dlr.sc.virsat.fdir.core.markov.scheduler.MarkovSchedulerTest;
@@ -36,7 +37,8 @@ import junit.framework.JUnit4TestAdapter;
 	MarkovSchedulerTest.class,
 	ModelCheckingResultTest.class,
 	MarkovAutomatonBuilderTest.class,
-	StronglyConnectedComponentTest.class
+	StronglyConnectedComponentTest.class,
+	StronglyConnectedComponentFinderTest.class
 	})
 
 public class AllTests {
